backend/requirements.txt
fastapi
uvicorn
pandas
numpy
scikit-learn
openai
python-multipart
pydantic
httpx
python-dotenv
google-generativeai
supabase
